Decouple Paparazzi from JUnit 4
